move: notice 페이지에서 컴포넌트로 파일 이동
